rename: PageTitle.jsx, RenteeInfo.jsx, SearchBar.jsx 디렉토리 변경
