+++ src/pages/MyPage.jsx
@@ -3,9 +3,9 @@ import { useState, useEffect } from "react";
 import UserAPI from "@/lib/api/UserAPI";
 
 import Body from "@/styles/Body";
-import PageTitle from "@/components/PageTitle";
+import PageTitle from "@/components/common/PageTitle";
 import MyPageHeader from "@/components/header/MyPageHeader";
-import RenteeInfo from "@/components/RenteeInfo";
+import RenteeInfo from "@/components/common/RenteeInfo";
 import Profile from "@/components/my-page/Profile";
 import Button from "@/components/my-page/Button";
